shell commit

+++ Shell-Scripts/Robo-Shell/ansible-files/shell-assignments/assignment-1.sh
@@ -13,13 +13,17 @@ N="\e[0m"
 # Use DAYS_TO_DELETE in your file deletion operation
 if [ $ACTION == "delete" ]; then
     FILES_TO_DELETE=$(find "$SOURCE_DIR" -type f -name "*.log" -mtime +"$DAYS_TO_DELETE")
-    echo "$FILES_TO_DELETE files to delete"
-    echo "User chose to delete the files older than $DAYS_TO_DELETE days"
-    if [ -n "$FILES_TO_DELETE" ]; then
-        while IFS= read -r line; do
-            echo -e "Deleting file: $line"
-            rm -rf "$line"
-        done <<<"$FILES_TO_DELETE"
+    if [ $FILES_TO_DELETE -gt 0 ]; then
+        echo "User chose to delete the files older than $DAYS_TO_DELETE days"
+        if [ -n "$FILES_TO_DELETE" ]; then
+            while IFS= read -r line; do
+                echo -e "Deleting file: $line"
+                rm -rf "$line"
+            done <<<"$FILES_TO_DELETE"
+        fi
+    else
+        echo "No file to delete for provided duration"
+        exit 1
     fi
 fi
